+++ CHANGELOG.md
@@ -11,6 +11,7 @@ as of 1.2.0.
 ### Changed
 
 - Making all requests through a single Session/connection pool
+- External link icon now bundled in-repo (#138)
 
 ## [1.2.1] - 2022-08-03